requirements.txt
# Python: 3.11–3.13 recommended (dbt-core may fail on 3.14 until upstream support lands).
dbt-core>=1.8.0,<2.0.0
dbt-postgres>=1.8.0,<2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
